Add a base class for native encoders
