+++ ext/src/AutobahnMsgpackLzfService.java
@@ -6,6 +6,7 @@ import org.jruby.RubyModule;
 import org.jruby.runtime.load.BasicLibraryService;
 
 import autobahn.encoder.MsgPackLzfEncoder;
+import autobahn.encoder.MsgPackEncoderBase;
 
 
 public class AutobahnMsgpackLzfService implements BasicLibraryService {
@@ -13,6 +14,7 @@ public class AutobahnMsgpackLzfService implements BasicLibraryService {
     RubyModule autobahnModule = runtime.getOrCreateModule("Autobahn");
     RubyClass encoderClass = autobahnModule.getClass("Encoder");
     RubyClass msgPackLzfClass = autobahnModule.defineClassUnder("MsgPackLzfEncoder", encoderClass, MsgPackLzfEncoder.ALLOCATOR);
+    msgPackLzfClass.defineAnnotatedMethods(MsgPackEncoderBase.class);
     msgPackLzfClass.defineAnnotatedMethods(MsgPackLzfEncoder.class);
     return true;
   }
